+++ index.js
@@ -33,7 +33,7 @@
     for (var k in source) {
       if (source.hasOwnProperty(k)) {
         if (overwrite !== true) {
-          assert(!target.hasOwnProperty(k), 'Cannot overwrite property %s', k);
+          assert(!target.hasOwnProperty(k), 'Invalid call to mixin(target, source, [overwrite]): cannot overwrite property %s of target object', k);
         }
         target[k] = source[k];
       }
@@ -63,7 +63,7 @@
   }
 
   function getFunctionName(f) {
-    assert(typeof f === 'function', 'Invalid argument `f` = `%s` supplied to `getFunctionName()`', f);
+    assert(typeof f === 'function', 'Invalid argument f %s supplied to getFunctionName(f)', f);
     return f.displayName || f.name || format('<function%s>', f.length);
   }
 
@@ -83,12 +83,12 @@
   };
 
   function getTypeName(type) {
-    assert(Type.is(type), 'Invalid argument `type` = `%s` supplied to `getTypeName()`', type);
+    assert(Type.is(type), 'Invalid argument type %s supplied to getTypeName(type)', type);
     return type.meta.name;
   }
 
   function blockNew(x, type) {
-    assert(!(x instanceof type), 'Operator `new` is forbidden for type `%s`', getTypeName(type));
+    assert(!(x instanceof type), 'Operator new is forbidden for type %s', getTypeName(type));
   }
 
   function shallowCopy(x) {
@@ -112,16 +112,19 @@
   }
 
   update.commands = {
-    '$apply': function (f, value) {
+
+    $apply: function (f, value) {
       assert(Func.is(f));
       return f(value);
     },
-    '$push': function (elements, arr) {
+
+    $push: function (elements, arr) {
       assert(Arr.is(elements));
       assert(Arr.is(arr));
       return arr.concat(elements);
     },
-    '$remove': function (keys, obj) {
+
+    $remove: function (keys, obj) {
       assert(Arr.is(keys));
       assert(Obj.is(obj));
       for (var i = 0, len = keys.length ; i < len ; i++ ) {
@@ -129,10 +132,12 @@
       }
       return obj;
     },
-    '$set': function (value) {
+
+    $set: function (value) {
       return value;
     },
-    '$splice': function (splices, arr) {
+
+    $splice: function (splices, arr) {
       assert(list(Arr).is(splices));
       assert(Arr.is(arr));
       return splices.reduce(function (acc, splice) {
@@ -140,7 +145,8 @@
         return acc;
       }, arr);
     },
-    '$swap': function (config, arr) {
+
+    $swap: function (config, arr) {
       assert(Obj.is(config));
       assert(Num.is(config.from));
       assert(Num.is(config.to));
@@ -150,18 +156,21 @@
       arr[config.from] = element;
       return arr;
     },
-    '$unshift': function (elements, arr) {
+
+    $unshift: function (elements, arr) {
       assert(Arr.is(elements));
       assert(Arr.is(arr));
       return elements.concat(arr);
     },
-    '$merge': function (obj, value) {
+
+    $merge: function (obj, value) {
       return mixin(mixin({}, value), obj, true);
     }
+
   };
 
-  function create(type, value, mut) {
-    return type.meta.kind === 'struct' ? new type(value, mut) : type(value, mut);
+  function create(type, value, mut, path) {
+    return type.meta.kind === 'struct' ? new type(value, mut, path) : type(value, mut, path);
   }
 
   //
@@ -170,12 +179,12 @@
 
   function irreducible(name, is) {
 
-    assert(typeof name === 'string', 'Invalid argument `name` = `%s` supplied to `irreducible()`', name);
-    assert(typeof is === 'function', 'Invalid argument `is` = `%s` supplied to `irreducible()`', is);
+    assert(typeof name === 'string', 'Invalid argument name %s supplied to irreducible(name, is) (expected a string)', name);
+    assert(typeof is === 'function', 'Invalid argument is %s supplied to irreducible(name, is)', is);
 
-    function Irreducible(value) {
+    function Irreducible(value, mut, path) {
       blockNew(this, Irreducible);
-      assert(is(value), 'Invalid argument `value` = `%s` supplied to irreducible type `%s`', value, name);
+      assert(is(value), 'Invalid value %j supplied to %s', value, path ? path.join('/') : name);
       return value;
     }
 
@@ -199,39 +208,39 @@
     return x === null || x === void 0;
   });
 
-  var Str = irreducible('Str', function (x) {
+  var Str = irreducible('String', function (x) {
     return typeof x === 'string';
   });
 
-  var Num = irreducible('Num', function (x) {
+  var Num = irreducible('Number', function (x) {
     return typeof x === 'number' && isFinite(x) && !isNaN(x);
   });
 
-  var Bool = irreducible('Bool', function (x) {
+  var Bool = irreducible('Boolean', function (x) {
     return x === true || x === false;
   });
 
-  var Arr = irreducible('Arr', function (x) {
+  var Arr = irreducible('Array', function (x) {
     return x instanceof Array;
   });
 
-  var Obj = irreducible('Obj', function (x) {
+  var Obj = irreducible('Object', function (x) {
     return !Nil.is(x) && typeof x === 'object' && !Arr.is(x);
   });
 
-  var Func = irreducible('Func', function (x) {
+  var Func = irreducible('Function', function (x) {
     return typeof x === 'function';
   });
 
-  var Err = irreducible('Err', function (x) {
+  var Err = irreducible('Error', function (x) {
     return x instanceof Error;
   });
 
-  var Re = irreducible('Re', function (x) {
+  var Re = irreducible('RegExp', function (x) {
     return x instanceof RegExp;
   });
 
-  var Dat = irreducible('Dat', function (x) {
+  var Dat = irreducible('Date', function (x) {
     return x instanceof Date;
   });
 
@@ -239,31 +248,42 @@
     return Func.is(x) && Obj.is(x.meta);
   });
 
-  function struct(props, name) {
-
-    assert(dict(Str, Type).is(props), 'Invalid argument `props` = `%s` supplied to `struct` combinator', props);
-    assert(maybe(Str).is(name), 'Invalid argument `name` = `%s` supplied to `struct` combinator', name);
-    name = name || format('{%s}', Object.keys(props).map(function (prop) {
+  function getDefaultStructName(props) {
+    return format('{%s}', Object.keys(props).map(function (prop) {
       return format('%s: %s', prop, getTypeName(props[prop]));
     }).join(', '));
+  }
 
-    function Struct(value, mut) {
+  function struct(props, name) {
+
+    assert(dict(Str, Type).is(props), 'Invalid argument props %j supplied to struct(props, [name]) combinator', props);
+    assert(maybe(Str).is(name), 'Invalid argument name %s supplied to struct(props, [name]) combinator (expected a string)', name);
+
+    name = name || getDefaultStructName(props);
+
+    function Struct(value, mut, path) {
+
       // makes Struct idempotent
       if (Struct.is(value)) {
         return value;
       }
-      assert(Obj.is(value), 'Invalid argument `value` = `%s` supplied to struct type `%s`', value, name);
+
+      path = path || [name];
+      assert(Obj.is(value), 'Invalid value %j supplied to %s (expected an object)', value, path.join('/'));
+
       // makes `new` optional
       if (!(this instanceof Struct)) {
         return new Struct(value, mut);
       }
+
       for (var k in props) {
         if (props.hasOwnProperty(k)) {
           var expected = props[k];
           var actual = value[k];
-          this[k] = create(expected, actual, mut);
+          this[k] = create(expected, actual, mut, path.concat(k + ': ' + getTypeName(expected)));
         }
       }
+
       if (mut !== true) {
         Object.freeze(this);
       }
@@ -298,21 +318,24 @@
     return Struct;
   }
 
+  function getDefaultUnionName(types) {
+    return types.map(getTypeName).join(' | ');
+  }
+
   function union(types, name) {
 
-    assert(list(Type).is(types), 'Invalid argument `types` = `%s` supplied to `union` combinator', types);
-    var len = types.length;
-    var defaultName = types.map(getTypeName).join(' | ');
-    assert(len >= 2, 'Invalid argument `types` = `%s` supplied to `union` combinator, provide at least two types', defaultName);
-    assert(maybe(Str).is(name), 'Invalid argument `name` = `%s` supplied to `union` combinator', name);
-    name = name || defaultName;
+    assert(list(Type).is(types), 'Invalid argument types %j supplied to union(types, [name]) combinator (expected a list of types)', types);
+    var lenCache = types.length;
+    assert(lenCache >= 2, 'Invalid argument types %j supplied to union(types, [name]) combinator (expected a list of at least two types)', types);
+    assert(maybe(Str).is(name), 'Invalid argument name %s supplied to union(types, [name]) combinator (expected a string)', name);
+    name = name || getDefaultUnionName(types);
 
-    function Union(value, mut) {
+    function Union(value, mut, path) {
       blockNew(this, Union);
-      assert(Func.is(Union.dispatch), 'Unimplemented `dispatch()` function for union type `%s`', name);
+      assert(Func.is(Union.dispatch), 'Unimplemented dispatch() function for the union type %s', name);
       var type = Union.dispatch(value);
-      assert(Type.is(type), 'The `dispatch()` function of union type `%s` returns no type constructor', name);
-      return create(type, value, mut);
+      assert(Type.is(type), 'The dispatch() function of the union type %s returns no type constructor', name);
+      return create(type, value, mut, path);
     }
 
     Union.meta = {
@@ -331,7 +354,7 @@
 
     // default dispatch implementation
     Union.dispatch = function (x) {
-      for (var i = 0 ; i < len ; i++ ) {
+      for (var i = 0 ; i < lenCache ; i++ ) {
         if (types[i].is(x)) {
           return types[i];
         }
@@ -341,19 +364,26 @@
     return Union;
   }
 
+  function getMaybeDefaultName(type) {
+    return '?' + getTypeName(type);
+  }
+
   function maybe(type, name) {
 
-    assert(Type.is(type), 'Invalid argument `type` = `%s` supplied to `maybe` combinator', type);
+    assert(Type.is(type), 'Invalid argument type %s supplied to maybe(type, [name]) combinator', type);
+
     // makes the combinator idempotent and handle Any, Nil
     if (type.meta.kind === 'maybe' || type === Any || type === Nil) {
       return type;
     }
-    assert(Nil.is(name) || Str.is(name), 'Invalid argument `name` = `%s` supplied to `maybe` combinator', name);
-    name = name || ('?' + getTypeName(type));
 
-    function Maybe(value, mut) {
+    assert(Nil.is(name) || Str.is(name), 'Invalid argument name %s supplied to maybe(type, [name]) combinator (expected a string)', name);
+
+    name = name || getMaybeDefaultName(type);
+
+    function Maybe(value, mut, path) {
       blockNew(this, Maybe);
-      return Nil.is(value) ? null : create(type, value, mut);
+      return Nil.is(value) ? null : create(type, value, mut, path);
     }
 
     Maybe.meta = {
@@ -371,16 +401,22 @@
     return Maybe;
   }
 
+  function getDefaultEnumsName(map) {
+    return Object.keys(map).map(function (k) { return JSON.stringify(k); }).join(' | ');
+  }
+
   function enums(map, name) {
 
-    assert(Obj.is(map), 'Invalid argument `map` = `%s` supplied to `enums` combinator', map);
-    assert(maybe(Str).is(name), 'Invalid argument `name` = `%s` supplied to `enums` combinator', name);
-    var keys = Object.keys(map); // cache enums
-    name = name || keys.map(function (k) { return JSON.stringify(k); }).join(' | ');
+    assert(Obj.is(map), 'Invalid argument map %j supplied to enums(map, [name]) combinator (expected an object)', map);
+    assert(maybe(Str).is(name), 'Invalid argument name %s supplied to enums(map, [name]) combinator (expected a string)', name);
 
-    function Enums(value) {
+    var keysCache = Object.keys(map); // cache enums
+
+    name = name || getDefaultEnumsName(map);
+
+    function Enums(value, mut, path) {
       blockNew(this, Enums);
-      assert(Enums.is(value), 'Invalid argument `value` = `%s` supplied to enums type `%s`, expected one of %j', value, name, keys);
+      assert(Enums.is(value), 'Invalid value %j supplied to %s (expected one of %j)', value, path ? path.join('/') : name, keysCache);
       return value;
     }
 
@@ -408,49 +444,65 @@
     return enums(value, name);
   };
 
+  function getDefaultTupleName(types) {
+    return format('[%s]', types.map(getTypeName).join(', '));
+  }
+
   function tuple(types, name) {
 
-    assert(list(Type).is(types), 'Invalid argument `types` = `%s` supplied to `tuple` combinator', types);
-    var len = types.length; // cache types length
-    assert(maybe(Str).is(name), 'Invalid argument `name` = `%s` supplied to `tuple` combinator', name);
-    name = name || format('[%s]', types.map(getTypeName).join(', '));
+    assert(list(Type).is(types), 'Invalid argument types %j supplied to tuple(types, [name]) combinator (expected a list of types)', types);
+
+    var lenCache = types.length; // cache types length
+
+    assert(maybe(Str).is(name), 'Invalid argument name %s supplied to tuple(types, [name]) combinator (expected a string)', name);
+
+    name = name || getDefaultTupleName(types);
 
     function isTuple(x) {
-      return types.every(function (type, i) {
-        return type.is(x[i]);
-      });
+      for (var i = 0 ; i < lenCache ; i++) {
+        if (!types[i].is(x[i])) {
+          return false;
+        }
+      }
+      return true;
     }
 
-    function Tuple(value, mut) {
-      assert(Arr.is(value) && value.length === len, 'Invalid argument `value` = `%s` supplied to tuple type `%s`, expected an `Arr` of length `%s`', value, name, len);
+    function Tuple(value, mut, path) {
+
+      path = path || [name];
+      assert(Arr.is(value) && value.length === lenCache, 'Invalid value %j supplied to %s (expected an array of length %s)', value, path.join('/'), lenCache);
+
       var frozen = (mut !== true);
+
       // makes Tuple idempotent
       if (isTuple(value) && Object.isFrozen(value) === frozen) {
         return value;
       }
+
       var arr = [];
-      for (var i = 0 ; i < len ; i++) {
+      for (var i = 0 ; i < lenCache ; i++) {
         var expected = types[i];
         var actual = value[i];
-        arr.push(create(expected, actual, mut));
+        arr.push(create(expected, actual, mut, path.concat(i + ': ' + getTypeName(expected))));
       }
+
       if (frozen) {
         Object.freeze(arr);
       }
+
       return arr;
     }
 
     Tuple.meta = {
       kind: 'tuple',
       types: types,
-      length: len,
       name: name
     };
 
     Tuple.displayName = name;
 
     Tuple.is = function (x) {
-      return Arr.is(x) && x.length === len && isTuple(x);
+      return Arr.is(x) && x.length === lenCache && isTuple(x);
     };
 
     Tuple.update = function (instance, spec) {
@@ -460,17 +512,23 @@
     return Tuple;
   }
 
+  function getDefaultSubtypeName(type, predicate) {
+    return format('{%s | %s}', getTypeName(type), getFunctionName(predicate));
+  }
+
   function subtype(type, predicate, name) {
 
-    assert(Type.is(type), 'Invalid argument `type` = `%s` supplied to `subtype` combinator', type);
-    assert(Func.is(predicate), 'Invalid argument `predicate` = `%s` supplied to `subtype` combinator', predicate);
-    assert(maybe(Str).is(name), 'Invalid argument `name` = `%s` supplied to `subtype` combinator', name);
-    name = name || format('{%s | %s}', getTypeName(type), getFunctionName(predicate));
+    assert(Type.is(type), 'Invalid argument type %s supplied to subtype(type, predicate, [name]) combinator', type);
+    assert(Func.is(predicate), 'Invalid argument predicate %s supplied to subtype(type, predicate, [name]) combinator', predicate);
+    assert(maybe(Str).is(name), 'Invalid argument name %s supplied to subtype(type, predicate, [name]) combinator (expected a string)', name);
 
-    function Subtype(value, mut) {
+    name = name || getDefaultSubtypeName(type, predicate);
+
+    function Subtype(value, mut, path) {
       blockNew(this, Subtype);
-      var x = create(type, value, mut);
-      assert(predicate(x), 'Invalid argument `value` = `%s` supplied to subtype type `%s`', value, name);
+      path = path || [name];
+      var x = create(type, value, mut, path);
+      assert(predicate(x), 'Invalid value %j supplied to %s', value, path.join('/'));
       return x;
     }
 
@@ -494,31 +552,45 @@
     return Subtype;
   }
 
+  function getDefaultListName(type) {
+    return format('Array<%s>', getTypeName(type));
+  }
+
   function list(type, name) {
 
-    assert(Type.is(type), 'Invalid argument `type` = `%s` supplied to `list` combinator', type);
-    assert(maybe(Str).is(name), 'Invalid argument `name` = `%s` supplied to `list` combinator', name);
-    name = name || format('Array<%s>', getTypeName(type));
+    assert(Type.is(type), 'Invalid argument type %s supplied to list(type, [name]) combinator', type);
+    assert(maybe(Str).is(name), 'Invalid argument name %s supplied to list(type, [name]) combinator (expected a string)', name);
+
+    name = name || getDefaultListName(type);
+
+    var typeNameCache = getTypeName(type);
 
     function isList(x) {
       return x.every(type.is);
     }
 
-    function List(value, mut) {
-      assert(Arr.is(value), 'Invalid argument `value` = `%s` supplied to list type `%s`', value, name);
+    function List(value, mut, path) {
+
+      path = path || [name];
+      assert(Arr.is(value), 'Invalid value %j supplied to %s (expected an array of %s)', value, path.join('/'), typeNameCache);
+
       var frozen = (mut !== true);
+
       // makes List idempotent
       if (isList(value) && Object.isFrozen(value) === frozen) {
         return value;
       }
+
       var arr = [];
       for (var i = 0, len = value.length ; i < len ; i++ ) {
         var actual = value[i];
-        arr.push(create(type, actual, mut));
+        arr.push(create(type, actual, mut, path.concat(i + ': ' + typeNameCache)));
       }
+
       if (frozen) {
         Object.freeze(arr);
       }
+
       return arr;
     }
 
@@ -541,40 +613,56 @@
     return List;
   }
 
+  function getDefaultDictName(domain, codomain) {
+    return format('{[key:%s]: %s}', getTypeName(domain), getTypeName(codomain));
+  }
+
   function dict(domain, codomain, name) {
 
-    assert(Type.is(domain), 'Invalid argument `domain` = `%s` supplied to `dict` combinator', domain);
-    assert(Type.is(codomain), 'Invalid argument `codomain` = `%s` supplied to `dict` combinator', codomain);
-    assert(maybe(Str).is(name), 'Invalid argument `name` = `%s` supplied to `dict` combinator', name);
-    name = name || format('{[key:%s]: %s}', getTypeName(domain), getTypeName(codomain));
+    assert(Type.is(domain), 'Invalid argument domain %s supplied to dict(domain, codomain, [name]) combinator', domain);
+    assert(Type.is(codomain), 'Invalid argument codomain %s supplied to dict(domain, codomain, [name]) combinator', codomain);
+    assert(maybe(Str).is(name), 'Invalid argument name %s supplied to dict(domain, codomain, [name]) combinator (expected a string)', name);
+
+    var codomainNameCache = getTypeName(codomain);
+
+    name = name || getDefaultDictName(domain, codomain);
 
     function isDict(x) {
       for (var k in x) {
         if (x.hasOwnProperty(k)) {
-          if (!domain.is(k) || !codomain.is(x[k])) { return false; }
+          if (!domain.is(k) || !codomain.is(x[k])) {
+            return false;
+          }
         }
       }
       return true;
     }
 
-    function Dict(value, mut) {
-      assert(Obj.is(value), 'Invalid argument `value` = `%s` supplied to dict type `%s`', value, name);
+    function Dict(value, mut, path) {
+
+      path = path || [name];
+      assert(Obj.is(value), 'Invalid value %j supplied to %s', value, path.join('/'));
+
       var frozen = (mut !== true);
+
       // makes Dict idempotent
       if (isDict(value) && Object.isFrozen(value) === frozen) {
         return value;
       }
+
       var obj = {};
       for (var k in value) {
         if (value.hasOwnProperty(k)) {
-          k = create(domain, k);
+          k = create(domain, k, mut, path);
           var actual = value[k];
-          obj[k] = create(codomain, actual, mut);
+          obj[k] = create(codomain, actual, mut, path.concat(k + ': ' + codomainNameCache));
         }
       }
+
       if (frozen) {
         Object.freeze(obj);
       }
+
       return obj;
     }
 
@@ -602,22 +690,29 @@
     return Func.is(f) && Obj.is(f.type);
   }
 
+  function getDefaultFuncName(domain, codomain) {
+    return format('(%s) => %s', domain.map(getTypeName).join(', '), getTypeName(codomain));
+  }
+
   function func(domain, codomain, name) {
 
     // handle handy syntax for unary functions
     domain = Arr.is(domain) ? domain : [domain];
-    assert(list(Type).is(domain), 'Invalid argument `domain` = `%s` supplied to `func` combinator', domain);
-    assert(Type.is(codomain), 'Invalid argument `codomain` = `%s` supplied to `func` combinator', codomain);
-    assert(maybe(Str).is(name), 'Invalid argument `name` = `%s` supplied to `func` combinator', name);
-    name = name || format('(%s) => %s', domain.map(getTypeName).join(', '), getTypeName(codomain));
-    var domainLen = domain.length; // cache the domain length
+
+    assert(list(Type).is(domain), 'Invalid argument domain %s supplied to func(domain, codomain, [name]) combinator', domain);
+    assert(Type.is(codomain), 'Invalid argument codomain %s supplied to func(domain, codomain, [name]) combinator', codomain);
+    assert(maybe(Str).is(name), 'Invalid argument name %s supplied to func(domain, codomain, [name]) combinator (expected a string)', name);
+
+    name = name || getDefaultFuncName(domain, codomain);
+
+    var domainLenCache = domain.length; // cache the domain length
 
     function Func(value) {
       // automatically instrument the function
       if (!isInstrumented(value)) {
         return Func.of(value);
       }
-      assert(Func.is(value), 'Invalid argument `value` = `%s` supplied to func type `%s`', value, name);
+      assert(Func.is(value), 'Invalid value %j supplied to %s', value, name);
       return value;
     }
 
@@ -632,7 +727,7 @@
 
     Func.is = function (x) {
       return isInstrumented(x) &&
-        x.type.domain.length === domainLen &&
+        x.type.domain.length === domainLenCache &&
         x.type.domain.every(function (type, i) {
           return type === domain[i];
         }) &&
@@ -651,9 +746,9 @@
       function fn() {
         var args = slice.call(arguments);
         var len = args.length;
-        var argsType = tuple(domain.slice(0, len));
+        var argsType = tuple(domain.slice(0, len), name + ' domain');
         args = argsType(args);
-        if (len === domainLen) {
+        if (len === domainLenCache) {
           /* jshint validthis: true */
           return create(codomain, f.apply(this, args));
         } else {
@@ -692,14 +787,23 @@
     Any: Any,
     Nil: Nil,
     Str: Str,
+    String: Str,
     Num: Num,
+    Number: Num,
     Bool: Bool,
+    Boolean: Bool,
     Arr: Arr,
+    Array: Arr,
     Obj: Obj,
+    Object: Obj,
     Func: Func,
+    Function: Func,
     Err: Err,
+    Error: Err,
     Re: Re,
+    RegExp: Re,
     Dat: Dat,
+    Date: Dat,
     Type: Type,
     irreducible: irreducible,
     struct: struct,